Improving the solution: adding items and related endpoints.
Now it's possible to add items one by one and also to retrieve all given
category id.

Some tests added and also error handling.

+++ science/src/main/resources/db/changelog/migrations/001.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `category` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(30) NOT NULL,
   PRIMARY KEY (`id`)
 );
